misc/HtmlTestSuite/results/test-0956.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>Root module scripts should always use UTF-8</title>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="../serve-with-content-type.py?fn=external-script-utf8.js&amp;ct=text/javascript&amp;dummy=1" type="module"/>
<script type="module">
test(function() {
  assert_equals(window.getSomeString(), "śćążź",
                'Should be decoded as UTF-8');
}, 'UTF-8 module script');
</script>

<script charset="windows-1250" src="../serve-with-content-type.py?fn=external-script-utf8.js&amp;ct=text/javascript&amp;dummy=2" type="module"/>
<script type="module">
test(function() {
  assert_equals(window.getSomeString(), "śćążź",
                'Should be decoded as UTF-8');
}, 'UTF-8 module script with wrong charset in attribute');
</script>

<script src="../serve-with-content-type.py?fn=external-script-utf8.js&amp;ct=text/javascript%3Bcharset=windows-1250&amp;dummy=3" type="module"/>
<script type="module">
test(function() {
  assert_equals(window.getSomeString(), "śćążź",
                'Should be decoded as UTF-8');
}, 'UTF-8 module script with wrong charset in Content-Type');
</script>

<script src="../serve-with-content-type.py?fn=external-script-windows1250.js&amp;ct=text/javascript&amp;dummy=1" type="module"/>
<script type="module">
test(function() {
  assert_not_equals(window.getSomeString(), "śćążź",
                    'Should be decoded as UTF-8');
}, 'Non-UTF-8 module script');
</script>

<script charset="windows-1250" src="../serve-with-content-type.py?fn=external-script-windows1250.js&amp;ct=text/javascript&amp;dummy=2" type="module"/>
<script type="module">
test(function() {
  assert_not_equals(window.getSomeString(), "śćążź",
                    'Should be decoded as UTF-8');
}, 'Non-UTF-8 module script with charset in attribute');
</script>

<script src="../serve-with-content-type.py?fn=external-script-windows1250.js&amp;ct=text/javascript%3Bcharset=windows-1250" type="module"/>
<script type="module">
test(function() {
  assert_not_equals(window.getSomeString(), "śćążź",
                    'Should be decoded as UTF-8');
}, 'Non-UTF-8 module script with charset in Content-Type');
</script>
</head><body/></html>
